add fees breakdwn on campaign preview
