+++ src/components/dashboard/forms/web3/previews.js
@@ -1,77 +1,8 @@
 import React from 'react'
-import {
-	Box,
-	List,
-	ListItem,
-	ListItemText,
-	ListSubheader,
-	ExpansionPanel,
-	ExpansionPanelSummary,
-	Typography,
-} from '@material-ui/core'
-import { ExpandMoreSharp as ExpandMoreIcon } from '@material-ui/icons'
+import { Box, ListItemText } from '@material-ui/core'
 
 import { PropRow } from 'components/common/dialog/content'
 import { t } from 'selectors'
-
-export const FeesBreakdown = ({ breakdownFormatted = {}, symbol }) => (
-	<Box p={1}>
-		<ExpansionPanel square={true} variant='outlined'>
-			<ExpansionPanelSummary
-				expandIcon={<ExpandMoreIcon />}
-				aria-controls='fees-breakdown'
-				id='fees-breakdown'
-			>
-				<Typography>{t('FEES_BREAKDOWN_ADVANCED')}</Typography>
-			</ExpansionPanelSummary>
-			<List
-				disablePadding
-				dense
-				subheader={
-					<ListSubheader component='div'>
-						{t('BD_TOTAL_FEE', {
-							args: [breakdownFormatted.feeAmount, symbol],
-						})}
-					</ListSubheader>
-				}
-			>
-				{!!breakdownFormatted.deployFee && (
-					<ListItem>
-						<ListItemText
-							primary={t('BD_DEPLOY_FEE', {
-								args: [breakdownFormatted.deployFee, symbol],
-							})}
-						/>
-					</ListItem>
-				)}
-				<ListItem>
-					<ListItemText
-						primary={t('BD_TXNS_FEE', {
-							args: [
-								breakdownFormatted.executeAction,
-								breakdownFormatted.txnsFee,
-								symbol,
-							],
-						})}
-					/>
-				</ListItem>
-				{!!breakdownFormatted.sweepTxnsCount && (
-					<ListItem>
-						<ListItemText
-							primary={t('BD_SWEEP_FEE', {
-								args: [
-									breakdownFormatted.sweepTxnsCount,
-									breakdownFormatted.sweepTxnsFeeAmount,
-									symbol,
-								],
-							})}
-						/>
-					</ListItem>
-				)}
-			</List>
-		</ExpansionPanel>
-	</Box>
-)
 
 export const IdentityWithdrawPreview = ({
 	withdrawTo,
